Patch for x86_64
Reported by Corinna Vinschen off-list:

expect has a bug.  It crashes in the openpty call on x86_64, because
it uses openpty like this:

	openpty(&master, &slave, master_name, 0, 0)

The 0 constants are obviously int (32 bit), but the openpty function
expects pointers (64 bit).  That would be no problem, if the prototype
of the openpty function would be available, but unfortunately the
pty_termios.c source neglects to include <pty.h> for some reason.
The end result is that the 0 is only written as 32 bit value to the
stack, so the openpty function gets a pointer with 32 random upper
bits and 32 lower 0-bits.